refactor: refactoring controller to old version
