--- src/controllers/index.ts
@@ -1,12 +1,0 @@
-import MeetingController from "./meetingController"
-import PhysicalRoomController from "./physicalRoomController"
-import VirtualRoomController from "./virtualRoomController"
-import UserController from "./userController"
-
-
-export {
-  MeetingController,
-  PhysicalRoomController,
-  VirtualRoomController,
-  UserController
-}
